# spec/api-openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionProbeBody": {
        "description": "探针请求：未保存的新连接直接传明文 api_key；已保存连接可只给 connection_id 复用密钥。",
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "connection_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Connection Id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "protocol": {
            "enum": [
              "openai",
              "anthropic"
            ],
            "title": "Protocol",
            "type": "string"
          }
        },
        "required": [
          "protocol",
          "base_url"
        ],
        "title": "ConnectionProbeBody",
        "type": "object"
      },
      "ConnectionTestOut": {
        "description": "模型连接探针：ping 联通结果（ok + 耗时；reply 为截断回显，可能为空）。",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "latency_ms": {
            "default": 0,
            "title": "Latency Ms",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          },
          "reply": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reply"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ConnectionTestOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ModelConnectionBody": {
        "properties": {
          "api_key": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Api Key"
          },
          "base_url": {
            "title": "Base Url",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "input_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Price"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "output_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Price"
          },
          "protocol": {
            "enum": [
              "openai",
              "anthropic"
            ],
            "title": "Protocol",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "protocol",
          "base_url",
          "model"
        ],
        "title": "ModelConnectionBody",
        "type": "object"
      },
      "ModelListOut": {
        "description": "模型连接探针：拉取可用模型 id（设置页「获取模型列表」；中转不支持 /models 时 ok=False）。",
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "models": {
            "items": {
              "type": "string"
            },
            "title": "Models",
            "type": "array"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "ModelListOut",
        "type": "object"
      },
...
      "ProjectSettingsOut": {
        "properties": {
          "genre_pack": {
            "additionalProperties": true,
            "title": "Genre Pack",
            "type": "object"
          },
          "model_connections": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Model Connections",
            "type": "array"
          },
          "model_routes": {
            "additionalProperties": true,
            "title": "Model Routes",
            "type": "object"
          },
          "skill_pack": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Pack"
          },
          "style_profile": {
            "additionalProperties": true,
            "title": "Style Profile",
            "type": "object"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "version"
        ],
        "title": "ProjectSettingsOut",
        "type": "object"
      },
...
      "SettingsBody": {
        "description": "两个字段都是「省略 = 保留现值，显式传（含 {} / []）= 整体替换」。\n\n原先 model_routes 省略会被当成空字典覆盖，只改连接的客户端会静默丢掉全部分角色配置。",
        "properties": {
          "model_connections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ModelConnectionBody"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Connections"
          },
          "model_routes": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "string"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Routes"
          }
        },
        "title": "SettingsBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Myink 内部 API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/projects/{project_id}/settings": {
      "get": {
        "operationId": "get_project_settings_internal_v1_projects__project_id__settings_get",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Project Settings",
        "tags": [
          "settings"
        ]
      },
      "put": {
        "operationId": "put_project_settings_internal_v1_projects__project_id__settings_put",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectSettingsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Put Project Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/settings/models": {
      "post": {
        "operationId": "list_connection_models_internal_v1_projects__project_id__settings_models_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionProbeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Connection Models",
        "tags": [
          "settings"
        ]
      }
    },
    "/internal/v1/projects/{project_id}/settings/test-connection": {
      "post": {
        "operationId": "test_model_connection_internal_v1_projects__project_id__settings_test_connection_post",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "required": true,
            "schema": {
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-Myink-User",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Myink-User"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionProbeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionTestOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Model Connection",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}